requirements.txt
Flask==2.2.5
Pillow==9.5.0
numpy==1.24.3
tensorflow==2.13.0  # Or your specific TensorFlow version
scikit-learn==1.3.0
opencv-python==<4.10.0.84>
